update view controller for admin

+++ workspace/TheBakerShop/src/main/java/com/iuh/spring/dao/CatergoryDAO.java
@@ -7,7 +7,7 @@ import com.iuh.spring.entity.Category;
 public interface CatergoryDAO {
 	public List<Category> getAllCategory();
 
-	public Category getCategoryById(int id);
+	public Category getCategoryById(long id);
 
 	public boolean insertCategory(Category category);
