+++ workspace/TheBakerShop/src/main/java/com/iuh/spring/service/CategoryService.java
@@ -7,7 +7,7 @@ import com.iuh.spring.entity.Category;
 public interface CategoryService {
 	public List<Category> getAllCategory();
 
-	public Category getCategoryById(int id);
+	public Category getCategoryById(long id);
 
 	public boolean insertCategory(Category category);
